Only remove infobar if it exists.
It's possible to call RemoveInfoBar with an invalid/inexistant ID,
in which case pInfoBar is NULL. There's no point then actually trying
to remove this inexistant infobar.

However, calling removeInfoBar on NULL doesn't actually cause any harm
(in it's current implementation) as it iterates over the infobar list
to find the relevant infobar for deletion - however it causes unnecessary
iterations over the whole list and potential window updates.

Change-Id: I64d68e95387d2743cdbec31a77d6296cd0578024